Search around date interval endpoints rather than from midpoint
I decided it was more useful to pick up a timestamp near the
endpoints (since thats what the reference date would be set to)
so that users who set overlapping intervals get the expected
behavior.  For instance, every 1 day, 2 days, and 6 days will
all choose to preserve the same backup.