Added visualization feature to track down repeatability bug
One of the main goals is to repeatably select the same
timestamps even if run from different reference dates.
However, this isn't working because sometimes one
interval will hold onto a date that is nearby one held
by a different interval, and then by gradually adjusting
the reference date they might then move to the one the
other had selected, and drop a timestamp that should be
kept.

The visualization mode was able to help track this down.